pyproject.toml
[project]
name = "excel-parser-bot"
version = "0.1.0"
description = ""
authors = [
    {name = "al-ov73",email = "klever1987@yandex.ru"}
]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "aiogram (>=3.20.0,<4.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "sqlite-utils (>=3.38,<4.0)",
    "openpyxl (>=3.1.5,<4.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
